requirements.txt
Flask
flask-cors
pandas
numpy
nltk
scikit-learn
python-dotenv
requests
